Fixed hashes for the new ROTL32

+++ crypto/ripemd/ripemd.odin
@@ -145,44 +145,44 @@ ripemd_128_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 16 {
 			alpha = a + (b ~ c ~ d) + x[RIPEMD_128_N0[i]];
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (bb & dd | cc &~ dd) + x[RIPEMD_128_N1[i]] + 0x50a28be6;
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd= dd, alpha, bb, cc;
 			i += 1;
 		}
 		for i < 32 {
 			alpha = a + (d ~ (b & (c~d))) + x[RIPEMD_128_N0[i]] + 0x5a827999;
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (dd ~ (bb | ~cc)) + x[RIPEMD_128_N1[i]] + 0x5c4dd124;
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd = dd, alpha, bb, cc;
 			i += 1;
 		}
 		for i < 48 {
 			alpha = a + (d ~ (b | ~c)) + x[RIPEMD_128_N0[i]] + 0x6ed9eba1;
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (dd ~ (bb & (cc~dd))) + x[RIPEMD_128_N1[i]] + 0x6d703ef3;
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd = dd, alpha, bb, cc;
 			i += 1;
 		}
 		for i < 64 {
 			alpha = a + (c ~ (d & (b~c))) + x[RIPEMD_128_N0[i]] + 0x8f1bbcdc;
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (bb ~ cc ~ dd) + x[RIPEMD_128_N1[i]];
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd = dd, alpha, bb, cc;
 			i += 1;
 		}
@@ -211,12 +211,12 @@ ripemd_160_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 16 {
 			alpha = a + (b ~ c ~ d) + x[RIPEMD_160_N0[i]];
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb ~ (cc | ~dd)) + x[RIPEMD_160_N1[i]] + 0x50a28be6;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -224,12 +224,12 @@ ripemd_160_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 32 {
 			alpha = a + (b&c | ~b&d) + x[RIPEMD_160_N0[i]] + 0x5a827999;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb&dd | cc&~dd) + x[RIPEMD_160_N1[i]] + 0x5c4dd124;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -237,12 +237,12 @@ ripemd_160_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 48 {
 			alpha = a + (b | ~c ~ d) + x[RIPEMD_160_N0[i]] + 0x6ed9eba1;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb | ~cc ~ dd) + x[RIPEMD_160_N1[i]] + 0x6d703ef3;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -250,12 +250,12 @@ ripemd_160_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 64 {
 			alpha = a + (b&d | c&~d) + x[RIPEMD_160_N0[i]] + 0x8f1bbcdc;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb&cc | ~bb&dd) + x[RIPEMD_160_N1[i]] + 0x7a6d76e9;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -263,12 +263,12 @@ ripemd_160_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 80 {
 			alpha = a + (b ~ (c | ~d)) + x[RIPEMD_160_N0[i]] + 0xa953fd4e;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb ~ cc ~ dd) + x[RIPEMD_160_N1[i]];
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -299,11 +299,11 @@ ripemd_256_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 16 {
 			alpha = a + (b ~ c ~ d) + x[RIPEMD_128_N0[i]];
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (bb & dd | cc &~ dd) + x[RIPEMD_128_N1[i]] + 0x50a28be6;
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd= dd, alpha, bb, cc;
 			i += 1;
 		}
@@ -313,11 +313,11 @@ ripemd_256_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 32 {
 			alpha = a + (d ~ (b & (c~d))) + x[RIPEMD_128_N0[i]] + 0x5a827999;
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (dd ~ (bb | ~cc)) + x[RIPEMD_128_N1[i]] + 0x5c4dd124;
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd = dd, alpha, bb, cc;
 			i += 1;
 		}
@@ -327,11 +327,11 @@ ripemd_256_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 48 {
 			alpha = a + (d ~ (b | ~c)) + x[RIPEMD_128_N0[i]] + 0x6ed9eba1;
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (dd ~ (bb & (cc~dd))) + x[RIPEMD_128_N1[i]] + 0x6d703ef3;
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd = dd, alpha, bb, cc;
 			i += 1;
 		}
@@ -341,11 +341,11 @@ ripemd_256_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 64 {
 			alpha = a + (c ~ (d & (b~c))) + x[RIPEMD_128_N0[i]] + 0x8f1bbcdc;
 			s := int(RIPEMD_128_R0[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			a, b, c, d = d, alpha, b, c;
 			alpha = aa + (bb ~ cc ~ dd) + x[RIPEMD_128_N1[i]];
 			s = int(RIPEMD_128_R1[i]);
-			alpha = ROTL32(alpha, u32(s));
+			alpha = ROTL32(alpha, s);
 			aa, bb, cc, dd = dd, alpha, bb, cc;
 			i += 1;
 		}
@@ -380,12 +380,12 @@ ripemd_320_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 16 {
 			alpha = a + (b ~ c ~ d) + x[RIPEMD_160_N0[i]];
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb ~ (cc | ~dd)) + x[RIPEMD_160_N1[i]] + 0x50a28be6;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -396,12 +396,12 @@ ripemd_320_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 32 {
 			alpha = a + (b&c | ~b&d) + x[RIPEMD_160_N0[i]] + 0x5a827999;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb&dd | cc&~dd) + x[RIPEMD_160_N1[i]] + 0x5c4dd124;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -412,12 +412,12 @@ ripemd_320_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 48 {
 			alpha = a + (b | ~c ~ d) + x[RIPEMD_160_N0[i]] + 0x6ed9eba1;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb | ~cc ~ dd) + x[RIPEMD_160_N1[i]] + 0x6d703ef3;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -428,12 +428,12 @@ ripemd_320_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 64 {
 			alpha = a + (b&d | c&~d) + x[RIPEMD_160_N0[i]] + 0x8f1bbcdc;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb&cc | ~bb&dd) + x[RIPEMD_160_N1[i]] + 0x7a6d76e9;
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
@@ -444,12 +444,12 @@ ripemd_320_block :: proc(ctx: ^$T, p: []byte) -> int {
 		for i < 80 {
 			alpha = a + (b ~ (c | ~d)) + x[RIPEMD_160_N0[i]] + 0xa953fd4e;
 			s := int(RIPEMD_160_R0[i]);
-			alpha = ROTL32(alpha, u32(s)) + e;
+			alpha = ROTL32(alpha, s) + e;
 			beta = ROTL32(c, 10);
 			a, b, c, d, e = e, alpha, b, beta, d;
 			alpha = aa + (bb ~ cc ~ dd) + x[RIPEMD_160_N1[i]];
 			s = int(RIPEMD_160_R1[i]);
-			alpha = ROTL32(alpha, u32(s)) + ee;
+			alpha = ROTL32(alpha, s) + ee;
 			beta = ROTL32(cc, 10);
 			aa, bb, cc, dd, ee = ee, alpha, bb, beta, dd;
 			i += 1;
